Reverted the relative positioning of morphs according to its owner.
The problem with that change is that it only took care of drawing but broke everything else (i.e. event handling, collision detection, positioning when adding a submorph, etc). I will implement the relative positioning correctly eventually but we need to take care of all these problems, otherwise it not useful.