Fixed mouse event bug regarding the cursor position.
By using the MouseEvent.x and y properties to track the cursor I was ignoring the offset from the parent element. Now I replaced that with offsetX and offsetY. However, these new properties seem to be incompatible between browsers so I might need to change this in the near future.